COH-523: Adding external acceptance tests to H1
This commit introduces no new behaviour to the H1 application, but does
add a number of acceptance tests. In order to run H1 in an acceptance
suite, the SequenceServer has been modified so that it now has start()
and stop() instance methods. The dependency version of jersey-common was
also increased.

A new acceptance test suite has been added to test simple POST and GET
requests to the server. This uses an embedded Zookeeper which is wrapped
as a JUnitRule/ExternalResource.

Finally, the existing unit test ZkClockTest was refactored to use the
new EmbeddedZookeeper.

Change-Id: Id2f577d5961a2a92dc8370dc80373d9478b92404