COH-479: Implementing batch query for sequence numbers
This commit introduces support for querying a number of keys using a
single http call. A json object representing the sequence for each key
is returned. Internally, this makes a number of calls to Zookeeper, if
any of them don't exist or cause an error, the whole query fails.

The query can be done with a simple GET using query params

curl http://localhost:9595/seq/?key=foo&key=bar

Change-Id: Ia948a4be782dc9d49ce3ab0a1eabb2252cd65e11