COH-540: Supporting sequence queries via HTTP POST
Previously, it was possible to query a number of sequences using a GET
with query parameters.

e.g. to get the sequences of foo and bar

curl http://localhost:9595/seq/?key=foo&key=bar

This commit allows the query to be made via a HTTP Form POST, allowing
the querying of large numbers of keys (JETTY currently throws an
exception if the size of the HTTP Headers (including the url) is greater
than 4Kb.

e.g. curl -d key=foo -d key=bar http://localhost:9595/seq/

Change-Id: Iaa1ec86490f3244e3c930074beefa0ea024ead5f