COH-526: H1 to return -1 for unknown sequences
H1 used to return an HTTP 404 or sequences that were missing from
Zookeeper. However, whenever we made a call to H1 and got this result we
translate it to be sequence -1. This commit changes the behaviour to be
an HTTP 200 repsonse with an entity of "-1".

Additionally a number of tweaks were made to the embedded zookeeper used
in the acceptance tests as there was a race between the test-cases.

Change-Id: Idf2fa0b8bc865081ac80e8748ae2a65450010900